add favicon and fix routing
